+++ hvad/contrib/restframework/__init__.py
@@ -1,0 +1,3 @@
+from .serializers import (TranslationsMixin,
+                          TranslatableModelSerializer,
+                          HyperlinkedTranslatableModelSerializer)